capsule-update: python and MACHINE fixes
Use native python correctly using python3native.bbclass and amend
custom PYTHONPATH to it. Then mark recipe compatible with only
machines from this layer since the recipe does nothing on other
ones. Buildin this recipe on other machines causes errors when
SRC_URI file is not found.

Note that UEFI_CAPSULE_CERT_FILE hack is still there and needs
to be cleaned up later. Pulling in files without going through
SRC_URI or DEPENDS is very, very bad practice in yocto.

+++ meta-trustedsubstrate/recipes-bsp/capsule-update/capsule-update_0.1.bb
@@ -1,16 +1,23 @@
 LICENSE = "MIT"
 LIC_FILES_CHKSUM = "file://${COREBASE}/meta/COPYING.MIT;md5=3da9cfbcb788c80a0384361b4de20420"
 
-inherit deploy
+inherit deploy python3native
 
+COMPATIBLE_MACHINE = "synquacer|rockpi4b|stm32mp157c-ev1|stm32mp157c-dk2|zynqmp-production|zynqmp-starter"
+
+SRC_URI = "file://${MACHINE}-images.json"
+
+DEPENDS = "\
+	ovmf-native \
+	python3-pyopenssl-native \
+	trusted-firmware-a \
+	u-boot \
+"
+
+# If certs not deployed
 ALLOW_EMPTY:${PN} = "1"
 
-DEPENDS += "ovmf-native u-boot trusted-firmware-a"
-DEPENDS += "python3-pyopenssl-native"
-
-SRC_URI  += "file://${MACHINE}-images.json"
-
-do_deploy() {
+do_compile() {
 	if [ -e "${UEFI_CAPSULE_CERT_FILE}" ]; then
 		mkdir -p uefi_capsule_certs
 		tar xpvfz ${UEFI_CAPSULE_CERT_FILE} -C uefi_capsule_certs
@@ -27,12 +34,16 @@ do_deploy() {
 		if [ -e ${DEPLOY_DIR_IMAGE}/ImageA.bin ]; then cp ${DEPLOY_DIR_IMAGE}/ImageA.bin . ; fi
 		if [ -e ${DEPLOY_DIR_IMAGE}/ImageB.bin ]; then cp ${DEPLOY_DIR_IMAGE}/ImageB.bin . ; fi
 
-		cp ${S}/../${MACHINE}-images.json  .
-		export PYTHONPATH="${STAGING_DIR_NATIVE}/usr/bin/edk2_basetools/BaseTools/Source/Python"
-		${HOSTTOOLS_DIR}/python3 ${STAGING_DIR_NATIVE}/usr/bin/edk2_basetools/BaseTools/Source/Python/Capsule/GenerateCapsule.py -j ${MACHINE}-images.json \
+		cp ${S}/../${MACHINE}-images.json .
+		export PYTHONPATH="${STAGING_DIR_NATIVE}/usr/bin/edk2_basetools/BaseTools/Source/Python:${PYTHONPATH}"
+		python3 ${STAGING_DIR_NATIVE}/usr/bin/edk2_basetools/BaseTools/Source/Python/Capsule/GenerateCapsule.py -j ${MACHINE}-images.json \
 			-e -o ${MACHINE}_fw.capsule \
 			--verbose
+	fi
+}
 
+do_deploy() {
+	if [ -e "${UEFI_CAPSULE_CERT_FILE}" ]; then
 		mkdir -p ${DEPLOYDIR}
 		cp ${MACHINE}_fw.capsule ${DEPLOYDIR}/
 	fi
@@ -41,4 +52,4 @@ do_deploy() {
 addtask deploy before do_build after do_compile
 
 CAPSULE_DEPENDS = " trusted-firmware-a:do_deploy u-boot:do_deploy"
-do_compile[deploy] .= "${CAPSULE_DEPENDS}"
+do_compile[depends] .= "${CAPSULE_DEPENDS}"
